% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbench-package.R
\docType{package}
\name{seqbench-package}
\alias{seqbench}
\alias{seqbench-package}
\title{seqbench: Anytime-Valid Sequential Benchmarking of Algorithms}
\description{
Sequential, anytime-valid comparison of algorithms on paired losses with a practical-equivalence margin. Confidence sequences for the mean paired difference, superiority/equivalence/continue decisions, cost accounting and auditable trajectories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/castlaboratory/seqbench}
  \item Report bugs at \url{https://github.com/castlaboratory/seqbench/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item André Leite \email{leite@castlab.org}
}

}
\keyword{internal}
